Remove redundant NameSpace and restricted implementations just use string type
