+++ immutable_test.go
@@ -1,0 +1,74 @@
+package uuid
+
+import (
+	"testing"
+)
+
+func TestMarshaller_MarshalBinary(t *testing.T) {
+	//id := Uuid(uuidBytes)
+	//u := Marshaller(id)
+	//bytes, err := u.MarshalBinary()
+	//assert.Nil(t, err, "There should be no error")
+	//assert.Equal(t, uuidBytes[:], bytes, "Byte should be the same")
+}
+
+func TestUuid_UnmarshalBinary(t *testing.T) {
+
+	//u := make(Uuid, length)
+	//
+	//err := u.UnmarshalBinary([]byte{1, 2, 3, 4, 5})
+	//
+	//assert.Equal(t, "uuid.Marshaller.UnmarshalBinary: invalid length", err.Error(), "Expect length error")
+	//
+	//err = u.UnmarshalBinary(uuidBytes[:])
+	//
+	//assert.Nil(t, err, "There should be no error but got %s", err)
+	//
+	//for k, v := range namespaces {
+	//	id, _ := Parse(v)
+	//	u = make(Uuid, length)
+	//	u.UnmarshalBinary(id.Bytes())
+	//
+	//	assert.Equal(t, id.Bytes(), mm.Bytes(), "The array id should equal the uuid id")
+	//	assert.Equal(t, k.Bytes(), mm.Bytes(), "The array id should equal the uuid id")
+	//}
+	//
+	//uu := uuid_array{}
+
+	//u = Marshaller(uu[:])
+	//err = u.UnmarshalBinary(uuidBytes[:])
+	//assert.Nil(t, err, "There should be no error but got %s", err)
+	//assert.Equal(t, uuidBytes[:], uu.Bytes(), "The array id should equal the uuid id")
+	//
+	//var ii UUID = V1()
+	//
+	//u = Marshaller(ii.(Uuid))
+	//err = u.UnmarshalBinary(uuidBytes[:])
+	//assert.Nil(t, err, "There should be no error but got %s", err)
+	//assert.Equal(t, uuidBytes[:], ii.Bytes(), "The array id should equal the uuid id")
+	//
+	//jj := new(uuid)
+	//
+	//u = Marshaller(Uuid(*jj))
+	//err = u.UnmarshalBinary(uuidBytes[:])
+	//assert.Error(t, err, "There should be an error")
+	//
+	//u = Marshaller(Uuid(Nil))
+	//err = u.UnmarshalBinary(uuidBytes[:])
+	//assert.Nil(t, err, "There should be no error but got %s", err)
+	//assert.NotEqual(t, uuidBytes[:], Nil.Bytes(), "The array id should equal the uuid id")
+	//
+	//u = Marshaller(Uuid(NameSpaceDNS))
+	//err = u.UnmarshalBinary(uuidBytes[:])
+	//assert.Nil(t, err, "There should be no error but got %s", err)
+	//assert.NotEqual(t, uuidBytes[:], NameSpaceDNS.Bytes(), "The array id should equal the uuid id")
+	//
+	//kk := V1()
+	//ll := kk.Unmodifiable()
+	//
+	//u = Marshaller(ll.(uuid))
+	//err = u.UnmarshalBinary(uuidBytes[:])
+	//assert.Nil(t, err, "There should be no error but got %s", err)
+	//assert.Equal(t, uuidBytes[:], ll.Bytes(), "The array id should equal the uuid id")
+
+}
